+++ src/SMAPI.ModBuildConfig.Analyzer.Tests/Mock/StardewValley/Farmer.cs
@@ -1,0 +1,11 @@
+// ReSharper disable CheckNamespace, InconsistentNaming -- matches Stardew Valley's code
+using System.Collections.Generic;
+
+namespace StardewValley
+{
+    /// <summary>A simplified version of Stardew Valley's <c>StardewValley.Farmer</c> class for unit testing.</summary>
+    internal class Farmer
+    {
+        public IDictionary<string, int[]> friendships;
+    }
+}