+++ src/SMAPI.ModBuildConfig.Analyzer.Tests/Mock/StardewValley/Object.cs
@@ -1,0 +1,12 @@
+// ReSharper disable CheckNamespace, InconsistentNaming -- matches Stardew Valley's code
+using Netcode;
+
+namespace StardewValley
+{
+    /// <summary>A simplified version of Stardew Valley's <c>StardewValley.Object</c> class for unit testing.</summary>
+    public class Object : Item
+    {
+        /// <summary>A net int field with an equivalent non-net property.</summary>
+        public NetInt type = new NetInt { Value = 42 };
+    }
+}