+++ src/SMAPI.ModBuildConfig.Analyzer/Properties/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Reflection;
+
+[assembly: AssemblyTitle("SMAPI.ModBuildConfig.Analyzer")]
+[assembly: AssemblyDescription("")]